add additional feedback on sequenceWriterAccess
